+++ lib/src/realm_class.dart
@@ -32,7 +32,7 @@ export 'package:realm_common/realm_common.dart'
     show Ignored, Indexed, MapTo, PrimaryKey, RealmError, RealmModel, RealmUnsupportedSetError, RealmCollectionType, RealmPropertyType;
 
 export "configuration.dart" show Configuration, RealmSchema, SchemaObject;
-export 'list.dart' show RealmList;
+export 'list.dart' show RealmList, RealmListOfObject;
 export 'realm_object.dart' show RealmException, RealmObject;
 export 'realm_property.dart';
 export 'results.dart' show RealmResults;