+++ CHANGELOG.md
@@ -4,6 +4,9 @@ vNext
 ### Changes
 * Primary key annotation no longer requires field to be final.
 
+### Enhancements
+* Support query on lists of realm objects
+
 
 0.2.0+alpha Release notes (2022-01-31)
 ==============================================================